Extract discovery query runner partial
